+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - add: creation time display
 - add: sort by creation time
+- add: randomly select
 # v0.2.0
 - improve: AddonNodeView
 - fix: WorkshopVpkAddon.FullVpkFilePath is sometimes not updated